requirements.txt
# ============================================
# Dependências Python - Projeto Trocas
# ============================================
# Execute: pip install -r requirements.txt

# Framework web Flask
Flask==3.0.0
# ORM para base de dados
Flask-SQLAlchemy==3.1.1
# Autenticação de utilizadores
Flask-Login==0.6.3
# Formulários web
Flask-WTF==1.2.1
# driver MySQL para Flask
Flask-MySQLdb==2.0.0
# Conector MySQL para Python
mysql-connector-python==8.2.0
# Biblioteca para hash de passwords
bcrypt==4.1.1
# Carregar variáveis de ambiente
python-dotenv==1.0.0
# Utilitários WSGI
Werkzeug==3.0.1
# Validação de emails
email-validator==2.1.0
# Processamento de imagens
Pillow==10.1.0
# Servidor WSGI para produção
gunicorn==21.2.0
# Driver PostgreSQL para Python
psycopg2-binary==2.9.9
